Moved san from ast to a dedicated separate module.
