+++ ast/meson.build
@@ -7,22 +7,5 @@ libast = static_library('ast',
                         dependencies: [ lexer_dep ],
                         install : true)
 
-ast_dep = declare_dependency(include_directories : '.',
+ast_dep = declare_dependency(include_directories : incdir,
                              link_with : libast)
-
-
-libvisitors = static_library('visitors', 
-                             sources : [
-                                 'impl/ast_printer.cpp',
-                                 'impl/identifier_resolver.cpp',
-                                 'impl/pretty_printer.cpp',
-                                 'impl/array_length_enforcer.cpp',
-                             ],
-                             cpp_args : riddle_cpp_args,
-                             c_args : riddle_c_args,
-                             link_args : riddle_link_args,
-                             dependencies : [ ast_dep, lexer_dep, fmt_dep, common_dep ],
-                             install : true)
-
-visitors_dep = declare_dependency(include_directories : '.',
-                               link_with : libvisitors)
