+++ meson.build
@@ -16,6 +16,7 @@ riddle_cpp_args = [
 riddle_c_args = []
 riddle_link_args = []
 
+
 if get_option('always-sanitize-address') or get_option('buildtype') == 'debug'
   warning('Compiling with address sanitizer. Performance may be affected.')
   flags = [ '-fsanitize=address' ]
@@ -24,12 +25,14 @@ if get_option('always-sanitize-address') or get_option('buildtype') == 'debug'
   riddle_link_args += flags
 endif
 
+
 fmt_dep = dependency('fmt', fallback : ['fmt', 'fmt_dep'])
 
 subdir('common')
 subdir('lexer')
 subdir('ast')
 subdir('parser')
+subdir('san')
 
 if get_option('build-tests')
   subdir('tests')
