+++ riddle/meson.build
@@ -3,4 +3,4 @@ executable('riddle',
           cpp_args : riddle_cpp_args,
           c_args : riddle_c_args,
           link_args : riddle_link_args,
-          dependencies : [ visitors_dep, parser_dep, ast_dep, fmt_dep, lexer_dep, common_dep ])
+          dependencies : [ san_dep, parser_dep, ast_dep, fmt_dep, lexer_dep, common_dep ])
