+++ san/impl/pretty_printer.cpp
@@ -1,6 +1,6 @@
 #include "ast.hpp"
 #include "fmt/core.h"
-#include "visitors.hpp"
+#include "san.hpp"
 #include <memory>
 #include <cassert>
